Install/uninstall plugins

+++ Sources/NewIcon/Core/TemplatePlugin.swift
@@ -70,45 +70,9 @@ struct TemplatePlugin {
         return .init(url: URL(fileURLWithPath: binPath).appendingPathComponent("libTemplate.dylib"))
     }
     
-    func cleanUp() throws {
-        guard installationURL.isTemporary else { return }
+    func cleanUp(forced: Bool = false) throws {
+        guard forced || installationURL.isTemporary else { return }
         try FileManager.default.removeItem(at: installationURL.url)
-    }
-}
-
-struct TemplateImage {
-    let url: URL
-    
-    func open<RenderTemplate>(
-        isTemplateSymbol: String,
-        renderTemplateSymbol: String,
-        renderTemplateType: RenderTemplate.Type
-    ) throws -> (types: [Any.Type], renderTemplate: RenderTemplate) {
-        guard let pluginHandle = dlopen(url.path, RTLD_NOW) else {
-            if let error = dlerror() {
-                throw "dlopen error: \(String(cString: error))"
-            } else {
-                throw "Unknown dlopen error"
-            }
-        }
-        
-        typealias IsTemplate = @convention(c) (Any) -> Bool
-        let isTemplate = try dlsym(pluginHandle, isTemplateSymbol)
-            .map { unsafeBitCast($0, to: (IsTemplate).self) }
-            .unwrapOrThrow("Plugin doesn't contain the \(isTemplateSymbol) entry point from TemplateSupport")
-        
-        let renderTemplate = try dlsym(pluginHandle, renderTemplateSymbol)
-            .map { unsafeBitCast($0, to: (RenderTemplate).self) }
-            .unwrapOrThrow("Plugin doesn't contain the \(renderTemplateSymbol) entry point from TemplateSupport")
-        
-        let templateTypes: [Any.Type] = try MachImage(name: url.path)
-            .unwrapOrThrow("Did not find plugin image after dlopen")
-            .symbols
-            .filter { $0.name.hasSuffix("VN") } // is struct type metadata? (see: https://github.com/apple/swift/blob/main/docs/ABI/Mangling.rst)
-            .map { unsafeBitCast($0.address, to: Any.Type.self) }
-            .filter(isTemplate)
-        
-        return (templateTypes, renderTemplate)
     }
 }
